# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EthioBiz API - Multilingual Business Discovery Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories with subcategories
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a single category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user (admin - suspend/role)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        suspended:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - role
        - suspended
        - createdAt
    UserUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [user, admin]
        suspended:
          type: boolean
    Category:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        nameEn:
          type: string
        nameAm:
          type: string
        nameOrm:
          type: string
        icon:
          type: ["string", "null"]
        parentId:
          type: ["integer", "null"]
        children:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        businessCount:
          type: integer
      required:
        - id
        - slug
        - nameEn
        - nameAm
        - nameOrm
        - businessCount
    CategoryInput:
      type: object
      properties:
        slug:
          type: string
        nameEn:
          type: string
        nameAm:
          type: string
        nameOrm:
          type: string
        icon:
          type: string
        parentId:
          type: ["integer", "null"]
      required:
        - slug
        - nameEn
        - nameAm
        - nameOrm
    CategoryUpdate:
      type: object
      properties:
        nameEn:
          type: string
        nameAm:
          type: string
        nameOrm:
          type: string
        icon:
          type: string
...