# lib/api-zod/src/generated/types/categoryUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EthioBiz API - Multilingual Business Discovery Platform
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryUpdate {
  nameEn?: string;
  nameAm?: string;
  nameOrm?: string;
  icon?: string;
}